Create sorting algorithm and rng generated arr
I created the mergesort method that can take an array of numbers,
and sort them from least to greatest, using recursion.To avoid having
to constantly type numbers for the array, I created a method that takes
a number, and maps them to an array. This can be called in the mergesort
method, to save time on typing out tests.